debugging

+++ Code/Models/Attention_seq2seq.py
@@ -270,7 +270,7 @@ class _Decoder(nn.Module):
         
         weighted = torch.bmm(attention, encoder_outputs)    # weighted = [batch size, 1, enc hid dim * 2]
         
-        weighted = weighted.permute(1, 0, 2).double() # weighted = [1, batch size, enc hid dim * 2]
+        weighted = weighted.permute(1, 0, 2) # weighted = [1, batch size, enc hid dim * 2]
 
         # print('embedded',embedded.size())
         rnn_input = torch.cat((embedded, weighted), dim=2).float()  # rnn_input = [1, batch size, (enc hid dim * 2) + dec_emb dim]
@@ -291,7 +291,7 @@ class _Decoder(nn.Module):
         weighted = weighted.squeeze(0)
         
         prediction = self.fc_out(
-            torch.cat((output, weighted, embedded), dim=1).float()
+            torch.cat((output, weighted, embedded), dim=1)
             )
 
         # prediction = [batch size, output dim]
